+++ src/middlewares/validateData.js
@@ -1,6 +1,6 @@
 import { assert } from "superstruct";
 import ownService from "../services/own-service.js";
-import { createShopStruct, updateShopStruct } from "../structs/shopStruct.js";
+import { createShopStruct, updateShopStruct } from "../structs/shop-struct.js";
 import { SignUpUser, SignInUser } from "../structs/user-struct.js";
 import shopService from "../services/shop-service.js";
 import userService from "../services/user-service.js";
@@ -8,7 +8,7 @@ import ownRepository from "../repositories/own-repository.js";
 import exchangeRepository from "../repositories/exchange-repository.js";
 import shopRepository from "../repositories/shop-repository.js";
 import { exchangeCardShopAndUserSelect } from "../repositories/selects/exchange-select.js";
-import { shopDetailSelect } from "../services/selects/shopSelect.js";
+import { shopDetailSelect } from "../services/selects/shop-select.js";
 
 export async function validateCreateShopData(req, res, next) {
   const userId = req.session.userId;
