+++ src/structs/shop-struct.js
